+++ src/pages/Sales/sale.tsx
@@ -1,12 +1,11 @@
-import "./Sale.css"
- import ProductsCardSale from '../../components/ProductsCardPage/ProductsCardSale';
-interface SaleProps {
-}
+import "./Sale.css";
+import ProductsCardSale from "../../components/ProductsCardPage/ProductsCardSale";
+interface SaleProps {}
 const Sale: React.FC<SaleProps> = () => {
   return (
     <>
-     <ProductsCardSale/>
-     </>
+      <ProductsCardSale />
+    </>
   );
 };
